Added "Show webcam" menu item to the testbed protocol.
svn path=/branches/kopete/0.12/kopete/; revision=532603

+++ libkopete/avdevice/videodevicepool.cpp
@@ -36,7 +36,7 @@ namespace Kopete {
 namespace AV {
 
 VideoDevicePool *VideoDevicePool::s_self = NULL;
-int VideoDevicePool::m_clients = 0;
+__u64 VideoDevicePool::m_clients = 0;
 
 VideoDevicePool* VideoDevicePool::self()
 {
