+++ libkopete/avdevice/videodevicepool.h
@@ -122,7 +122,7 @@ protected:
 private:
 	VideoDevicePool();
 	static VideoDevicePool* s_self;
-	static int m_clients; // Number of instances
+	static __u64 m_clients; // Number of instances
 };
 
 }
